+++ src/Ordenacao/MetricaOrdenacao.java
@@ -1,30 +1,30 @@
 package Ordenacao;
 
 public class MetricaOrdenacao {
-	int comparacao;
-	int troca;
+	long comparacao;
+	long troca;
 	
-	public MetricaOrdenacao(int comparacao, int troca) {
+	public MetricaOrdenacao(long comparacao, long troca) {
 		this.comparacao = comparacao;
 		this.troca = troca;
 	}
- 
-	public int getComparacao() {
+
+	public long getComparacao() {
 		return comparacao;
 	}
- 
-	public void setComparacao(int comparacao) {
+
+	public void setComparacao(long comparacao) {
 		this.comparacao = comparacao;
 	}
- 
-	public int getTroca() {
+
+	public long getTroca() {
 		return troca;
 	}
- 
-	public void setTroca(int troca) {
+
+	public void setTroca(long troca) {
 		this.troca = troca;
 	}
- 
+
 	@Override
 	public String toString() {
 		StringBuilder builder = new StringBuilder();
@@ -36,4 +36,3 @@ public class MetricaOrdenacao {
 		return builder.toString();
 	}
 }
-
